boilerplate for new stats page
