+++ webapp/package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "@types/react-dom": "^18.0.9",
+    "canvasjs": "^1.8.3",
     "pong-engine": "^1.0.0",
     "react": "^18.2.0",
     "react-confirm-alert": "^3.0.6",
@@ -11,8 +12,8 @@
     "react-router-dom": "^6.4.5",
     "react-scripts": "^5.0.1",
     "socket.io-client": "^4.5.4",
-    "web-vitals": "^3.1.0",
-    "transcendence-shared": "^1.0.0"
+    "transcendence-shared": "^1.0.0",
+    "web-vitals": "^3.1.0"
   },
   "scripts": {
     "start": "react-scripts start",
