+++ webapp/src/index.tsx
@@ -30,6 +30,8 @@ import {
   EnableTwoFactorAuthPage,
   ValidateTwoFactorAuthPage,
   GameRulesPage,
+  UserMatchHistoryPage,
+  UserStatsPage,
 } from './pages';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import {
@@ -65,7 +67,6 @@ import { UserBlocklistProvider } from './shared/context/UserBlocklistContext';
 import { UserFriendProvider } from './shared/context/UserFriendContext';
 import RequireAdmin from './shared/components/RequireAdmin/RequireAdmin';
 import { GamePairingProvider } from './shared/context/GamePairingContext';
-import UserStatsPage from './pages/UserStatsPage/UserStatsPage';
 
 const container = document.getElementById('root');
 const root = createRoot(container!);
@@ -179,6 +180,14 @@ function AppRoutes() {
         element={
           <RequireAuth>
             <UserPage />
+          </RequireAuth>
+        }
+      />
+      <Route
+        path={`${USER_URL}/:username/history`}
+        element={
+          <RequireAuth>
+            <UserMatchHistoryPage />
           </RequireAuth>
         }
       />
